# Refactors serializers/views for speed in enrollment APIs, adds flag to designate requirements as electives or not