# What are the relevant tickets?

Closes #1695 
Closes #1700 

# Description (What does it do?)

For #1695 -
- Adds `electives_flag` - if `depth=2` in a ProgramRequirement node, this means the child course nodes are electives (and thus eliminates the need to depend on the title for this distinction). 

For #1700 - 
- Optimizes queries in a few places to make them pull data with less nested queries
- Culls some of the recursion that was being used; we were walking the requirements tree to find things and there were more efficient ways to do this (e.g. using `path` in combination with the new `electives_flag`). 

Testing this locally dropped the response time for `/api/program_enrollments` from ~11-12s to ~1.2s (with debug set to True) or ~3-5s to ~1.2s (with debug set to False). This was running thousands of queries before; it should be much less now.

# How can this be tested?

* Automated tests should work.
* All functionality should work as expected - the main thing that uses the program enrollments API is the My Programs tab and program drawer in the dashboard. In addition, certificate generation should work as normal.